+++ Kernel/FileSystem/SysFS/Subsystems/Kernel/Configuration/BooleanVariable.cpp
@@ -1,10 +1,10 @@
 /*
- * Copyright (c) 2022, Liav A. <liavalb@hotmail.co.il>
+ * Copyright (c) 2022-2023, Liav A. <liavalb@hotmail.co.il>
  *
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include <Kernel/FileSystem/SysFS/Subsystems/Kernel/Variables/BooleanVariable.h>
+#include <Kernel/FileSystem/SysFS/Subsystems/Kernel/Configuration/BooleanVariable.h>
 #include <Kernel/Sections.h>
 #include <Kernel/Tasks/Process.h>
