+++ Kernel/FileSystem/SysFS/Subsystems/Kernel/Configuration/StringVariable.cpp
@@ -1,10 +1,10 @@
 /*
- * Copyright (c) 2022, Liav A. <liavalb@hotmail.co.il>
+ * Copyright (c) 2022-2023, Liav A. <liavalb@hotmail.co.il>
  *
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include <Kernel/FileSystem/SysFS/Subsystems/Kernel/Variables/StringVariable.h>
+#include <Kernel/FileSystem/SysFS/Subsystems/Kernel/Configuration/StringVariable.h>
 #include <Kernel/Sections.h>
 #include <Kernel/Tasks/Process.h>
